fix(db): correct impact_data.event_id type note + scene_age_days type consistency
Live introspection shows impact_data.event_id is actually UUID on Neon, not
TEXT as the app-level DDL string in agents/impact/services/db.py claims (the
DDL was aspirational/stale, never what actually ran) - schema.sql now states
the real type. Also aligns 0002's scene_age_days re-statement to DOUBLE
PRECISION to match 0001's actual type (agent.py's own inline ALTER used
DOUBLE PRECISION for fractional day counts), and marks CLAUDE.md's
"schema.sql is stale" warning as resolved now that it's synced with live
Neon.

+++ shared/db/migrations/0002_durable_evidence_trail.sql
@@ -26,10 +26,14 @@ ALTER TABLE satellite_results ADD COLUMN IF NOT EXISTS coverage_percent REAL;
 ALTER TABLE satellite_results ADD COLUMN IF NOT EXISTS coverage_status TEXT;
 -- scene_age_days already added in 0001_baseline_drift.sql (it was confirmed
 -- ABSENT from live Neon entirely, not merely unwritten — see that file and
--- root CLAUDE.md). Re-stated here with IF NOT EXISTS for idempotency /
--- readability of this migration as a self-contained diff of the durable-
--- evidence-trail feature.
-ALTER TABLE satellite_results ADD COLUMN IF NOT EXISTS scene_age_days INTEGER;
+-- root CLAUDE.md) as DOUBLE PRECISION, matching agent.py's own
+-- _persist_satellite_result inline ALTER (fractional days, e.g. 1.5).
+-- Re-stated here with IF NOT EXISTS for idempotency / readability of this
+-- migration as a self-contained diff of the durable-evidence-trail feature;
+-- this statement is a no-op given 0001 already ran, so the type here MUST
+-- match 0001's (DOUBLE PRECISION) or the IF NOT EXISTS guard makes this
+-- line's own type declaration irrelevant/misleading.
+ALTER TABLE satellite_results ADD COLUMN IF NOT EXISTS scene_age_days DOUBLE PRECISION;
 ALTER TABLE satellite_results ADD COLUMN IF NOT EXISTS index_calibrated BOOLEAN;
 ALTER TABLE satellite_results ADD COLUMN IF NOT EXISTS index_units TEXT;
 ALTER TABLE satellite_results ADD COLUMN IF NOT EXISTS selection_reason TEXT;
